refactor(crypto): split key_protocol.rs into pure verification + async signing (#1446)
Prerequisite for scp-protocol crate extraction. Separates pure sync
protocol logic (verification, HPKE, wire types, NonceDedup) from async
signing operations (epoch advance, key request, block notification).
Promotes hkdf_derive_key, aes128gcm_encrypt, aes128gcm_decrypt to pub
for cross-module access.
